GSDX:  - Use float instead of int for the video framerate.  - Use 59.94 instead of 60 for the ntsc framerate. This replaces a previous hackfix with a better one, but it's still not ideal. The ideal solution for the video encoding side would be to use an actual fraction (60000/1001) and pass this fraction to the encoder. The ideal solution for the gsdx side would be to deduce the real framerate from the timing parameters.
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@4925 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ plugins/GSdx/GSCapture.h
@@ -44,7 +44,7 @@ public:
 	GSCapture();
 	virtual ~GSCapture();
 
-	bool BeginCapture(int fps);
+	bool BeginCapture(float fps);
 	bool DeliverFrame(const void* bits, int pitch, bool rgba);
 	bool EndCapture();
